requirements.txt
asgiref==3.5.2
Django==4.1
django-taggit==3.0.0
sqlparse==0.4.2
tzdata==2022.2
